@@ -1,36 +1,41 @@
 name: just_another_workout_timer
 description: Just Another Workout Timer
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-version: 1.7.0+20210518
+version: 2.0.0+20210830
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_tts:
-  path_provider:
-  shared_preferences:
-  fluttertoast:
-  preferences:
-  url_launcher:
-  wakelock:
-  soundpool:
-  intl_utils:
-  scrollable_positioned_list:
+  flutter_tts: 3.2.2
+  path_provider: 2.0.2
+  flutter_file_dialog: 2.1.1
+  pref: 2.4.0
+  fluttertoast: 8.0.8
+  prefs: 3.3.0
+  url_launcher: 6.0.9
+  wakelock: 0.5.3+3
+  soundpool: 2.1.0
+  intl_utils: 2.4.0
+  json_annotation: 4.1.0
+  scrollable_positioned_list: 0.2.0-nullsafety.0
   flutter_localizations:
     sdk: flutter
+  uuid: 3.0.4
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  #integration_test:
-  flutter_launcher_icons:
-  flutter_oss_licenses:
-  effective_dart:
+  flutter_launcher_icons: ^0.9.2
+  flutter_oss_licenses: ^1.0.1
+  flutter_native_splash: ^1.2.1
+  effective_dart: ^1.3.2
+  json_serializable: ^5.0.0
+  build_runner: ^2.1.1
+
+dependency_overrides:
+  meta: ^1.7.0
+  analyzer: ^2.0.0
 
 flutter:
   uses-material-design: true
@@ -45,5 +50,17 @@ flutter_icons:
   android: "launcher_icon"
   ios: false
   image_path: "assets/ic_launcher.png"
+
 flutter_intl:
   enabled: true
+  localizely:
+    project_id: 6e3724ce-b0c2-459a-a840-311eb82328da
+
+flutter_native_splash:
+  color: "#303030"
+  image: "assets/ic_launcher.png"
+  android_gravity: center
+  android: true
+  ios: false
+  web: false
+  android12: true
