@@ -10,7 +10,7 @@ import 'workout_runner.dart';
 
 /// Main screen
 class HomePage extends StatefulWidget {
-  HomePage({Key key}) : super(key: key);
+  HomePage() : super();
 
   @override
   _HomePageState createState() => _HomePageState();
@@ -18,76 +18,40 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage> {
   List<Workout> workouts = [];
-  IconData _sortIcon = Icons.sort_by_alpha;
-  String _sortMode = 'alpha';
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       _loadWorkouts();
     });
   }
 
   /// load all workouts from disk and populate list
   _loadWorkouts() async {
-    var data = await getAllWorkouts();
-
-    setState(() {
-      workouts = _sortWorkouts(data);
-    });
+    getAllWorkouts().then((value) => setState(() {
+          workouts = value;
+          _saveSorting();
+        }));
   }
 
-  _updateSortMode() {
-    switch (_sortMode) {
-      case 'alpha':
-        {
-          _sortIcon = Icons.sort;
-          _sortMode = 'duration';
-        }
-        break;
-
-      case 'duration':
-        {
-          _sortIcon = Icons.sort_by_alpha;
-          _sortMode = 'alpha';
-        }
-        break;
-
-      default:
-        _sortIcon = Icons.sort_by_alpha;
+  _saveSorting() {
+    for (var workout in workouts.asMap().entries) {
+      workout.value.position = workout.key;
+      writeWorkout(workout.value);
     }
-    _loadWorkouts();
-  }
-
-  _sortWorkouts(List<Workout> workouts) {
-    var sortedWorkouts = List<Workout>.from(workouts);
-    switch (_sortMode) {
-      case 'alpha':
-        {
-          sortedWorkouts.sort((w1, w2) => w1.title.compareTo(w2.title));
-        }
-        break;
-
-      case 'duration':
-        {
-          sortedWorkouts.sort((w1, w2) => w2.duration.compareTo(w1.duration));
-        }
-        break;
-    }
-    return sortedWorkouts;
   }
 
   /// aks user if they want to delete a workout
   _showDeleteDialog(BuildContext context, Workout workout) {
     // set up the buttons
-    Widget cancelButton = FlatButton(
+    Widget cancelButton = TextButton(
       child: Text(S.of(context).cancel),
       onPressed: () {
         Navigator.of(context).pop();
       },
     );
-    Widget continueButton = FlatButton(
+    Widget continueButton = TextButton(
       child: Text(S.of(context).delete),
       onPressed: () {
         deleteWorkout(workout.title);
@@ -111,19 +75,29 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  Widget _buildWorkoutList() => ListView.builder(
-        itemBuilder: (context, index) {
-          if (index < workouts.length) {
-            return _buildWorkoutItem(workouts[index]);
-          } else {
-            return null;
+  Widget _buildWorkoutList() => ReorderableListView(
+        onReorder: (oldIndex, newIndex) {
+          if (oldIndex < newIndex) {
+            newIndex -= 1;
           }
+          setState(() {
+            var workout = workouts.removeAt(oldIndex);
+            workouts.insert(newIndex, workout);
+          });
+          _saveSorting();
         },
+        children: workouts.map(_buildWorkoutItem).toList(),
       );
 
   Widget _buildWorkoutItem(Workout workout) => Card(
-          child: Row(
+      key: Key(workout.toJson().toString()),
+      child: Row(
         children: [
+          Container(
+            padding: EdgeInsets.symmetric(horizontal: 4),
+            child: ReorderableDragStartListener(
+                child: Icon(Icons.drag_handle), index: workout.position),
+          ),
           Expanded(
             child: ListTile(
               title: Text(workout.title),
@@ -161,7 +135,12 @@ class _HomePageState extends State<HomePage> {
               tooltip: S.of(context).deleteWorkout,
               onPressed: () {
                 _showDeleteDialog(context, workout);
-              })
+              }),
+          IconButton(
+              onPressed: () {
+                exportWorkout(workout.title);
+              },
+              icon: Icon(Icons.save_alt))
         ],
       ));
 
@@ -170,11 +149,6 @@ class _HomePageState extends State<HomePage> {
         appBar: AppBar(
           title: Text(S.of(context).workouts),
           actions: [
-            IconButton(
-                icon: Icon(_sortIcon),
-                onPressed: () {
-                  setState(_updateSortMode);
-                }),
             IconButton(
                 icon: Icon(Icons.settings),
                 onPressed: () {
@@ -194,7 +168,7 @@ class _HomePageState extends State<HomePage> {
               context,
               MaterialPageRoute(
                 builder: (context) => BuilderPage(
-                  workout: Workout.empty(),
+                  workout: Workout(),
                   newWorkout: true,
                 ),
               ),
