@@ -1,19 +1,22 @@
 import 'package:flutter/services.dart';
-import 'package:preferences/preference_service.dart';
+import 'package:prefs/prefs.dart';
 import 'package:soundpool/soundpool.dart';
 
 // ignore: avoid_classes_with_only_static_members
 class SoundHelper {
-  static Soundpool _soundpool;
-  static int _beepLowId;
-  static int _beepHighId;
-  static int _tickId;
+  static final Soundpool _soundpool = Soundpool.fromOptions(
+      options: SoundpoolOptions(
+    streamType: StreamType.music,
+  ));
+  static late int _beepLowId;
+  static late int _beepHighId;
+  static late int _tickId;
+
   static bool useSound = false;
 
   static Future<void> loadSounds() async {
-    _soundpool ??= Soundpool(streamType: StreamType.music);
     await _loadSounds();
-    useSound = PrefService.getString('sound') == 'beep' ?? true;
+    useSound = Prefs.getString('sound') == 'beep';
   }
 
   static Future<void> _loadSounds() async {
@@ -34,7 +37,7 @@ class SoundHelper {
   }
 
   static void playBeepTick() {
-    if (PrefService.getBool('ticks')) _soundpool.play(_tickId);
+    if (Prefs.getBool('ticks')) _soundpool.play(_tickId);
   }
 
   static void playDouble() {
@@ -42,6 +45,16 @@ class SoundHelper {
       _soundpool.play(_beepLowId);
       Future.delayed(Duration(milliseconds: 200))
           .then((value) => _soundpool.play(_beepLowId));
+    }
+  }
+
+  static void playTriple() {
+    if (useSound) {
+      _soundpool.play(_beepHighId);
+      Future.delayed(Duration(milliseconds: 150))
+          .then((value) => _soundpool.play(_beepHighId))
+          .then((value) => Future.delayed(Duration(milliseconds: 150))
+              .then((value) => _soundpool.play(_beepHighId)));
     }
   }
 }
