@@ -12,8 +12,7 @@ class BuilderPage extends StatefulWidget {
   final Workout workout;
   final bool newWorkout;
 
-  BuilderPage({Key key, @required this.workout, @required this.newWorkout})
-      : super(key: key);
+  BuilderPage({required this.workout, required this.newWorkout}) : super();
 
   @override
   _BuilderPageState createState() =>
@@ -22,13 +21,13 @@ class BuilderPage extends StatefulWidget {
 
 /// page allowing a user to create a workout
 class _BuilderPageState extends State<BuilderPage> {
-  Workout _workout;
-  String _oldTitle;
-  bool _newWorkout;
+  late Workout _workout;
+  late String _oldTitle;
+  late bool _newWorkout;
   bool _dirty = false;
   int _lastDuration = 30;
 
-  _BuilderPageState(Workout workout, {bool newWorkout}) {
+  _BuilderPageState(Workout workout, {required bool newWorkout}) {
     _workout = workout;
     _oldTitle = _workout.title;
     _newWorkout = newWorkout;
@@ -36,7 +35,7 @@ class _BuilderPageState extends State<BuilderPage> {
 
   void _addSet() {
     setState(() {
-      _workout.sets.add(Set.empty());
+      _workout.sets.add(Set(exercises: []));
       _dirty = true;
     });
   }
@@ -74,9 +73,11 @@ class _BuilderPageState extends State<BuilderPage> {
               ));
       return;
     }
+
     setState(() {
       _workout.cleanUp();
     });
+
     if ((_newWorkout && await workoutExists(_workout.title)) ||
         (!_newWorkout &&
             _oldTitle != _workout.title &&
@@ -86,13 +87,13 @@ class _BuilderPageState extends State<BuilderPage> {
           builder: (context) => AlertDialog(
                 content: Text(S.of(context).overwriteExistingWorkout),
                 actions: <Widget>[
-                  FlatButton(
+                  TextButton(
                     child: Text(S.of(context).no),
                     onPressed: () {
                       Navigator.of(context).pop();
                     },
                   ),
-                  FlatButton(
+                  TextButton(
                     child: Text(S.of(context).yes),
                     onPressed: () async {
                       await deleteWorkout(_oldTitle);
@@ -131,213 +132,198 @@ class _BuilderPageState extends State<BuilderPage> {
     });
   }
 
-  Widget _buildSetList() => ListView.builder(
-        itemBuilder: (context, index) {
-          if (index < _workout.sets.length) {
-            return _buildSetItem(_workout.sets[index], index);
-          } else {
-            return null;
+  Widget _buildSetList() => ReorderableListView(
+        onReorder: (oldIndex, newIndex) {
+          if (oldIndex < newIndex) {
+            newIndex -= 1;
           }
+          setState(() {
+            var set = _workout.sets.removeAt(oldIndex);
+            _workout.sets.insert(newIndex, set);
+          });
         },
+        children: _workout.sets
+            .asMap()
+            .map((index, set) => MapEntry(index, _buildSetItem(set, index)))
+            .values
+            .toList(),
       );
 
-  Widget _buildSetItem(Set set, int index) => Card(
-      key: Key(set.toRawJson()),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              IconButton(
-                  icon: Icon(Icons.delete),
-                  tooltip: S.of(context).deleteSet,
-                  onPressed: () {
-                    _deleteSet(index);
-                  }),
-              Expanded(
-                child: ListTile(
-                  title: Text(
-                    S.of(context).setIndex(_workout.sets.indexOf(set) + 1),
-                  ),
-                  subtitle: Text('${Utils.formatSeconds(set.duration)}'),
-                ),
-              ),
-              Text(S.of(context).repetitions),
-              NumberStepper(
-                  lowerLimit: 1,
-                  upperLimit: 99,
-                  largeSteps: false,
-                  formatNumber: false,
-                  value: set.repetitions,
-                  valueChanged: (repetitions) {
-                    setState(() {
-                      set.repetitions = repetitions;
-                      _dirty = true;
-                    });
-                  })
-            ],
-          ),
-          _buildExerciseList(set, index),
-          ButtonBar(
-            alignment: MainAxisAlignment.spaceBetween,
+  Widget _buildSetItem(Set set, int index) =>
+      ReorderableDelayedDragStartListener(
+          index: index,
+          key: Key(set.id),
+          child: Card(
+              child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Row(
                 children: [
-                  IconButton(
-                    icon: Icon(Icons.fitness_center),
-                    tooltip: S.of(context).addExercise,
-                    onPressed: () {
-                      _addExercise(index, false);
-                    },
+                  Container(
+                    padding: EdgeInsets.symmetric(horizontal: 4),
+                    child: ReorderableDragStartListener(
+                        child: Icon(Icons.drag_handle), index: index),
                   ),
-                  IconButton(
-                    icon: Icon(Icons.pause_circle_filled),
-                    tooltip: S.of(context).addRest,
-                    onPressed: () {
-                      _addExercise(index, true);
-                    },
+                  Expanded(
+                    child: ListTile(
+                      title: Text(
+                        S.of(context).setIndex(_workout.sets.indexOf(set) + 1),
+                      ),
+                      subtitle: Text('${Utils.formatSeconds(set.duration)}'),
+                    ),
                   ),
+                  Text(S.of(context).repetitions),
+                  NumberStepper(
+                      lowerLimit: 1,
+                      upperLimit: 99,
+                      largeSteps: false,
+                      formatNumber: false,
+                      value: set.repetitions,
+                      valueChanged: (repetitions) {
+                        setState(() {
+                          set.repetitions = repetitions;
+                          _dirty = true;
+                        });
+                      })
                 ],
               ),
-              Row(
+              _buildExerciseList(set, index),
+              ButtonBar(
+                alignment: MainAxisAlignment.spaceBetween,
                 children: [
-                  IconButton(
-                    icon: Icon(Icons.copy),
-                    tooltip: S.of(context).duplicate,
-                    onPressed: () => _duplicateSet(index),
+                  Row(
+                    children: [
+                      IconButton(
+                        icon: Icon(Icons.fitness_center),
+                        tooltip: S.of(context).addExercise,
+                        onPressed: () {
+                          _addExercise(index, false);
+                        },
+                      ),
+                      IconButton(
+                        icon: Icon(Icons.pause_circle_filled),
+                        tooltip: S.of(context).addRest,
+                        onPressed: () {
+                          _addExercise(index, true);
+                        },
+                      ),
+                    ],
                   ),
-                  IconButton(
-                    icon: Icon(Icons.arrow_upward),
-                    tooltip: S.of(context).moveSetUp,
-                    onPressed: index - 1 >= 0
-                        ? () {
-                            setState(() {
-                              _workout.moveSet(index, moveUp: true);
-                              _dirty = true;
-                            });
-                          }
-                        : null,
+                  Row(
+                    children: [
+                      IconButton(
+                          icon: Icon(Icons.delete),
+                          tooltip: S.of(context).deleteSet,
+                          onPressed: () {
+                            _deleteSet(index);
+                          }),
+                      IconButton(
+                        icon: Icon(Icons.copy),
+                        tooltip: S.of(context).duplicate,
+                        onPressed: () => _duplicateSet(index),
+                      ),
+                    ],
                   ),
-                  IconButton(
-                    icon: Icon(Icons.arrow_downward),
-                    tooltip: S.of(context).moveSetDown,
-                    onPressed: index + 1 < _workout.sets.length
-                        ? () {
-                            setState(() {
-                              _workout.moveSet(index, moveUp: false);
-                              _dirty = true;
-                            });
-                          }
-                        : null,
-                  )
                 ],
-              ),
+              )
             ],
-          )
-        ],
-      ));
+          )));
 
-  Widget _buildExerciseList(Set set, int setIndex) => ListView.builder(
-      shrinkWrap: true,
-      primary: false,
-      itemBuilder: (context, index) {
-        if (index < _workout.sets[setIndex].exercises.length) {
-          return _buildExerciseItem(setIndex, index, set.exercises[index].name);
-        } else {
-          return null;
-        }
-      });
+  Widget _buildExerciseList(Set set, int setIndex) => ReorderableListView(
+        shrinkWrap: true,
+        primary: false,
+        onReorder: (oldIndex, newIndex) {
+          if (oldIndex < newIndex) {
+            newIndex -= 1;
+          }
+          setState(() {
+            var ex = _workout.sets[setIndex].exercises.removeAt(oldIndex);
+            _workout.sets[setIndex].exercises.insert(newIndex, ex);
+          });
+        },
+        children: set.exercises
+            .asMap()
+            .keys
+            .map((index) =>
+                _buildExerciseItem(setIndex, index, set.exercises[index].name))
+            .toList(),
+      );
 
-  Widget _buildExerciseItem(int setIndex, int exIndex, String name) => Card(
-        color: Theme.of(context).backgroundColor,
-        child: Row(
-          key: Key(_workout.sets[setIndex].exercises[exIndex].toRawJson()),
-          children: [
-            IconButton(
-              icon: Icon(Icons.delete),
-              tooltip: S.of(context).deleteExercise,
-              onPressed: () {
-                _deleteExercise(setIndex, exIndex);
-              },
-            ),
-            Expanded(
-                flex: 1,
-                child: TextFormField(
-                  initialValue: name,
-                  maxLength: 30,
-                  maxLengthEnforced: true,
-                  inputFormatters: [LengthLimitingTextInputFormatter(30)],
-                  maxLines: 1,
-                  decoration: InputDecoration(
-                    labelText: S.of(context).exercise,
-                  ),
-                  onChanged: (text) {
-                    _workout.sets[setIndex].exercises[exIndex].name = text;
-                    _dirty = true;
-                  },
-                )),
-            Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.center,
+  Widget _buildExerciseItem(int setIndex, int exIndex, String name) =>
+      ReorderableDelayedDragStartListener(
+          index: exIndex,
+          key: Key(_workout.sets[setIndex].exercises[exIndex].id),
+          child: Card(
+            color: Theme.of(context).backgroundColor,
+            child: Row(
+              key: Key(_workout.sets[setIndex].exercises[exIndex].id),
               children: [
-                NumberStepper(
-                  lowerLimit: 1,
-                  upperLimit: 999,
-                  largeSteps: true,
-                  formatNumber: true,
-                  value: _workout.sets[setIndex].exercises[exIndex].duration,
-                  valueChanged: (duration) {
-                    setState(() {
-                      _workout.sets[setIndex].exercises[exIndex].duration =
-                          duration;
-                      _dirty = true;
-                      _lastDuration = duration;
-                    });
-                  },
+                Container(
+                  padding: EdgeInsets.symmetric(horizontal: 6),
+                  child: ReorderableDragStartListener(
+                      child: Icon(Icons.drag_handle), index: exIndex),
                 ),
-                Row(
+                Expanded(
+                    child: Column(
                   children: [
-                    IconButton(
-                      icon: Icon(Icons.arrow_upward),
-                      padding: EdgeInsets.zero,
-                      tooltip: S.of(context).moveExerciseUp,
-                      onPressed: exIndex - 1 >= 0
-                          ? () {
-                              setState(() {
-                                _workout.sets[setIndex]
-                                    .moveExercise(exIndex, moveUp: true);
-                                _dirty = true;
-                              });
-                            }
-                          : null,
+                    TextFormField(
+                      initialValue: name,
+                      maxLength: 30,
+                      inputFormatters: [LengthLimitingTextInputFormatter(30)],
+                      maxLines: 1,
+                      decoration: InputDecoration(
+                        labelText: S.of(context).exercise,
+                      ),
+                      onChanged: (text) {
+                        _workout.sets[setIndex].exercises[exIndex].name = text;
+                        _dirty = true;
+                      },
                     ),
-                    IconButton(
-                      icon: Icon(Icons.arrow_downward),
-                      padding: EdgeInsets.zero,
-                      tooltip: S.of(context).moveExerciseDown,
-                      onPressed:
-                          exIndex + 1 < _workout.sets[setIndex].exercises.length
-                              ? () {
-                                  setState(() {
-                                    _workout.sets[setIndex]
-                                        .moveExercise(exIndex, moveUp: false);
-                                    _dirty = true;
-                                  });
-                                }
-                              : null,
-                    ),
-                    IconButton(
-                      icon: Icon(Icons.copy),
-                      tooltip: S.of(context).duplicate,
-                      onPressed: () => _duplicateExercise(setIndex, exIndex),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: [
+                        IconButton(
+                          icon: Icon(Icons.delete),
+                          tooltip: S.of(context).deleteExercise,
+                          onPressed: () {
+                            _deleteExercise(setIndex, exIndex);
+                          },
+                        ),
+                        IconButton(
+                          icon: Icon(Icons.copy),
+                          tooltip: S.of(context).duplicate,
+                          onPressed: () =>
+                              _duplicateExercise(setIndex, exIndex),
+                        ),
+                      ],
+                    )
+                  ],
+                )),
+                Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    NumberStepper(
+                      lowerLimit: 1,
+                      upperLimit: 999,
+                      largeSteps: true,
+                      formatNumber: true,
+                      value:
+                          _workout.sets[setIndex].exercises[exIndex].duration,
+                      valueChanged: (duration) {
+                        setState(() {
+                          _workout.sets[setIndex].exercises[exIndex].duration =
+                              duration;
+                          _dirty = true;
+                          _lastDuration = duration;
+                        });
+                      },
                     ),
                   ],
-                )
+                ),
               ],
             ),
-          ],
-        ),
-      );
+          ));
 
   @override
   Widget build(BuildContext context) => WillPopScope(
@@ -351,13 +337,13 @@ class _BuilderPageState extends State<BuilderPage> {
               builder: (context) => AlertDialog(
                     content: Text(S.of(context).exitCheck),
                     actions: <Widget>[
-                      FlatButton(
+                      TextButton(
                         child: Text(S.of(context).no),
                         onPressed: () {
                           Navigator.of(context).pop(false);
                         },
                       ),
-                      FlatButton(
+                      TextButton(
                         child: Text(S.of(context).yesExit),
                         onPressed: () {
                           Navigator.of(context).pop(true);
@@ -366,14 +352,13 @@ class _BuilderPageState extends State<BuilderPage> {
                     ],
                   ));
 
-          return value;
+          return value!;
         },
         child: Scaffold(
           appBar: AppBar(
             title: TextFormField(
                 initialValue: _workout.title,
                 maxLength: 30,
-                maxLengthEnforced: true,
                 inputFormatters: [LengthLimitingTextInputFormatter(30)],
                 maxLines: 1,
                 onChanged: (name) {
