@@ -1,40 +1,45 @@
 import 'dart:async';
 
 import 'package:flutter_tts/flutter_tts.dart';
-import 'package:preferences/preference_service.dart';
+import 'package:prefs/prefs.dart';
 
-// ignore: avoid_classes_with_only_static_members
 /// handles everything related to TTS
 class TTSHelper {
-  static FlutterTts flutterTts;
+  static FlutterTts flutterTts = FlutterTts();
 
   static bool available = true;
 
   /// enable/disable TTS output
   static bool useTTS = true;
 
+  static bool isTalking = false;
+
   static Future<void> init() async {
     flutterTts = FlutterTts();
 
-    useTTS = PrefService.getString('sound') == 'tts' ?? true;
+    useTTS = Prefs.getString('sound') == 'tts';
 
-    var ttsLang = PrefService.getString('tts_lang') ?? 'en-US';
-    if (ttsLang.endsWith('*')) ttsLang = ttsLang.replaceAll('*', '');
-    PrefService.setString('tts_lang', ttsLang);
+    var ttsLang = Prefs.getString('tts_lang', 'en-US');
+    await Prefs.setString('tts_lang', ttsLang);
+
+    flutterTts.setStartHandler(() {
+      isTalking = true;
+    });
+    flutterTts.setCompletionHandler(() {
+      isTalking = false;
+    });
 
     try {
       await flutterTts
           .setLanguage(ttsLang)
           .timeout(Duration(seconds: 1))
           .then((_) async {
-        await flutterTts.setSpeechRate(1.0);
         await flutterTts.setVolume(1.0);
-        await flutterTts.setPitch(1.0);
       });
     } on TimeoutException {
       available = false;
       useTTS = false;
-      PrefService.setString('sound', 'beep');
+      await Prefs.setString('sound', 'beep');
       return;
     }
   }
